# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Recipe Service API
  version: 1.0.0
paths:
  /authors:
    post:
      summary: Create a new author
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAuthorRequest'
      responses:
        '201':
          description: Author created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '400':
          description: Bad Request - Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /authors/{author_id}/recipes:
    get:
      summary: Get a list of recipes by a specific author with pagination
      parameters:
        - name: author_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          description: Maximum number of recipes to return (max 10)
          required: false
          schema:
            type: integer
            default: 10
            maximum: 10
        - name: offset
          in: query
          description: Number of recipes to skip before starting to collect the result set
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: A list of recipes by the author
          content:
            application/json:
              schema:
                type: object
                properties:
                  recipes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Recipe'
                  total:
                    type: integer
                    description: Total number of recipes available for the author
        '400':
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Author not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /recipes:
    get:
      summary: Get a list of recipes with pagination
      parameters:
        - name: limit
          in: query
          description: Maximum number of recipes to return (max 10)
          required: false
          schema:
            type: integer
            default: 10
            maximum: 10
        - name: offset
          in: query
          description: Number of recipes to skip before starting to collect the result set
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: A list of recipes
          content:
            application/json:
              schema:
                type: object
                properties:
                  recipes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Recipe'
                  total:
                    type: integer
                    description: Total number of recipes available
        '400':
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecipeRequest'
      responses:
        '201':
          description: Recipe created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '400':
          description: Bad Request - Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    CreateAuthorRequest:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
        bio:
          type: string
      required:
        - login
        - password

    CreateRecipeRequest:
      type: object
      properties:
        author_id:
          type: string
          format: uuid
        title:
          type: string
        instructions:
          type: string
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/IngredientQuantity'
      required:
        - author_id
        - title
        - instructions
        - ingredients

    IngredientQuantity:
      type: object
      properties:
        ingredient_id:
          type: string
          format: uuid
        quantity:
          type: number
          format: float
      required:
        - ingredient_id
        - quantity

    Recipe:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        instructions:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - title
        - instructions
        - created_at

    Author:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        bio:
          type: string
      required:
        - id
        - user_id

    ErrorResponse:
      type: object
      properties:
        error:
          type: string